Remove the OTP check in the response
The nonce check should be sufficient to verify that the response matches
the request. Alternative keyboard layouts are supported by the Yubikey
validation api, but the OTP value that is returned is normalized into
the QUERTY keyboard character set. The OTP equality check prevents
dvorak users from validating their yubikey.